% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{customBs4DashSidebar}
\alias{customBs4DashSidebar}
\title{Custom bs4DashSidebar function, changing target option in
brandTag to "_blank"}
\usage{
customBs4DashSidebar(
  ...,
  inputId = NULL,
  disable = FALSE,
  title = NULL,
  skin = "dark",
  status = "primary",
  brandColor = NULL,
  url = NULL,
  src = NULL,
  elevation = 4,
  opacity = 0.8,
  expand_on_hover = TRUE
)
}
\description{
Custom bs4DashSidebar function, changing target option in
brandTag to "_blank"
}
