% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{global_data}
\alias{global_data}
\title{ECDC data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 56471 rows and 19 columns.
}
\source{
\url{https://opendata.ecdc.europa.eu}
}
\usage{
global_data
}
\description{
Data from the European Centre for Disease Control.
}
\keyword{datasets}
