% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interventions_data}
\alias{interventions_data}
\title{Countryinterventions data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 23923 rows and 11 columns.
}
\source{
\url{https://data.humdata.org/}
}
\usage{
interventions_data
}
\description{
Countryinterventions data from The Humanitarian Data Exchange
}
\keyword{datasets}
