% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irish_county_data}
\alias{irish_county_data}
\title{Irish county data}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 6032 rows and 9 columns.
}
\source{
\url{http://opendata-geohive.hub.arcgis.com}
}
\usage{
irish_county_data
}
\description{
Irish county data from Irish government.
}
\keyword{datasets}
