% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irish_data}
\alias{irish_data}
\title{Irish government data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 290 rows and 5 columns.
}
\source{
\url{https://opendata.arcgis.com}
}
\usage{
irish_data
}
\description{
Irish data from Irish government.
}
\keyword{datasets}
