% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-themes.R
\name{use_css_bs4Dash_distill}
\alias{use_css_bs4Dash_distill}
\title{Use bs4Dash distill theme css}
\usage{
use_css_bs4Dash_distill(path = NULL)
}
\arguments{
\item{path}{Dir where the css file will be copied to.
Default to inst/app/www/custom.css.}
}
\description{
Use bs4Dash distill theme css
}
