% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-server.R
\name{value_box_counts}
\alias{value_box_counts}
\title{Creates the Ireland value boxes in the Summary tab}
\usage{
value_box_counts(tab, variable, title, icon, status = "success")
}
\description{
Creates the Ireland value boxes in the Summary tab
}
